# Use Hash#fetch with block