--- a/lib/overcommit/hook/pre_commit/chamber_verification.rb
+++ b/lib/overcommit/hook/pre_commit/chamber_verification.rb
@@ -5,7 +5,7 @@ module Overcommit::Hook::PreCommit
   # rubocop:disable Metrics/CyclomaticComplexity, Metrics/PerceivedComplexity
   class ChamberVerification < Base
     def run
-      approver_name  = config.fetch('approver_name', 'your approver')
+      approver_name  = config.fetch('approver_name') { 'your approver' }
       approver_email = config['approver_email'] ? " (#{config['approver_email']})" : nil
 
       result = execute(command)