--- a/lib/overcommit/utils.rb
+++ b/lib/overcommit/utils.rb
@@ -172,7 +172,7 @@ module Overcommit
         end
 
         result =
-          if (splittable_args = options.fetch(:args, [])).any?
+          if (splittable_args = options.fetch(:args) { [] }).any?
             debug(initial_args.join(' ') + " ... (#{splittable_args.length} splittable args)")
             Overcommit::CommandSplitter.execute(initial_args, options)
           else