--- a/spec/support/shell_helpers.rb
+++ b/spec/support/shell_helpers.rb
@@ -21,10 +21,10 @@ module ShellHelpers
   # @param options [Hash]
   # @raise [Timeout::TimeoutError] timeout has elapsed before condition holds
   def wait_until(options = {})
-    Timeout.timeout(options.fetch(:timeout, 1)) do
+    Timeout.timeout(options.fetch(:timeout) { 1 }) do
       loop do
         return if yield
-        sleep options.fetch(:check_interval, 0.1)
+        sleep options.fetch(:check_interval) { 0.1 }
       end
     end
   end