According to `fasterer`, "Hash#fetch with second argument is slower than Hash#fetch with block"

This provides a very small improvement but since this gem is used with various workflows I feel that any speed improvement is a win.

Before

```
Finished in 1 minute 12.6 seconds (files took 0.49573 seconds to load)
1343 examples, 0 failures
```

After

```
Finished in 1 minute 12.47 seconds (files took 0.48929 seconds to load)
1343 examples, 0 failures
```