@sds 2018-11-05
The difference in timing of a tenth of a second for tests that take over a minute to run are likely due to random jitter rather than actual performance benefit. 

While I could see lazy evaluation bringing benefits with respect to fewer objects allocated, this seems like a sacrifice in readability for questionable benefit.

With all that said—I will merge if you get the CI build passing. Thanks!