@HarlemSquirrel 2018-11-06
I have mixed feelings about this as well. The readability suffers most with blocks containing an empty hash.